// src/RickAndMorty/interfaces.tsx
export interface IState {
    episodes: [];
    favorites: Array<any>;
  }
  
export interface IActions {
      type: string,
      payload: any,
  }

  export interface IEpisode {
    airdate: string;
    airstamp: string;
    airtime: string;
    id: number;
    image: { medium: string; original: string };
    name: string;
    number: number;
    runtime: number;
    season: number;
    summary: string;
    type: string;
    url: string;
  }